ko biet da xong chua

+++ src/main/java/com/shopkeeper/lam/models/BookInfo.java
@@ -8,7 +8,7 @@ public final class BookInfo extends ProductInfo{
         productType=ProductType.BOOK;
     }
 
-    public BookInfo(String title, long productInfoId, String description, Category category, LocalDate releaseDate, double currentSalePrice, Publisher publisher, ArrayList<Person> contributors, double rating, ArrayList<String> award) {
+    public BookInfo(String title, int productInfoId, String description, Category category, LocalDate releaseDate, double currentSalePrice, Publisher publisher, ArrayList<Person> contributors, double rating, ArrayList<String> award) {
         super(title, productInfoId, description, category, releaseDate, currentSalePrice, publisher, contributors, rating, award);
         this.productType=ProductType.BOOK;
     }
