+++ src/main/java/com/shopkeeper/lam/models/FilmInfo.java
@@ -8,7 +8,7 @@ public final class FilmInfo extends ProductInfo{
         this.productType=ProductType.FILM;
     }
 
-    public FilmInfo(String title, long productInfoId, String description, Category category,  LocalDate releaseDate, double currentSalePrice, Publisher publisher, ArrayList<Person> contributors, double rating, ArrayList<String> award) {
+    public FilmInfo(String title, int productInfoId, String description, Category category,  LocalDate releaseDate, double currentSalePrice, Publisher publisher, ArrayList<Person> contributors, double rating, ArrayList<String> award) {
         super(title, productInfoId, description, category, releaseDate, currentSalePrice, publisher, contributors, rating, award);
         this.productType=ProductType.FILM;
     }
