+++ src/main/java/com/shopkeeper/lam/models/SongInfo.java
@@ -8,7 +8,7 @@ public final class SongInfo extends ProductInfo {
         this.productType=ProductType.SONG;
     }
 
-    public SongInfo(String title, long productInfoId, String description, Category category, LocalDate releaseDate, double currentSalePrice, Publisher publisher, ArrayList<Person> contributors, double rating, ArrayList<String> award) {
+    public SongInfo(String title, int productInfoId, String description, Category category, LocalDate releaseDate, double currentSalePrice, Publisher publisher, ArrayList<Person> contributors, double rating, ArrayList<String> award) {
         super(title, productInfoId, description, category, releaseDate, currentSalePrice, publisher, contributors, rating, award);
         this.productType=ProductType.SONG;
     }
